+++ ipyvizzu.py
@@ -28,32 +28,14 @@ class DisplayTemplate:
         </script>
         """
 
-    _NEW_CHART = """
+    INIT = _SCRIPT.format(
+        """
             let myVizzu_{c_id} = document.createElement("div");
             myVizzu_{c_id}.style.cssText = "width: {div_width}; height: {div_height};";
             let chart_{c_id} = import("{vizzu}").then(Vizzu => new Vizzu.default(myVizzu_{c_id}).initializing);
-            """
-
-    INIT = {
-        DisplayTarget.BEGIN: _SCRIPT.format(
-            f"""
-            {_NEW_CHART}
-            myVizzu_{{id}}.parentNode.insertBefore(myVizzu_{{c_id}}, myVizzu_{{id}});
-            """
-        ),
-        DisplayTarget.ACTUAL: _SCRIPT.format(
-            f"""
-            myVizzu_{{id}}.parentNode.parentNode.style.display = "none";
-            {_NEW_CHART}
-            """
-        ),
-        DisplayTarget.END: _SCRIPT.format(
-            f"""
-            myVizzu_{{id}}.parentNode.parentNode.style.display = "none";
-            {_NEW_CHART}
-            """
-        ),
-    }
+            myVizzu_{id}.parentNode.insertBefore(myVizzu_{c_id}, myVizzu_{id});
+            """  # pylint: disable=line-too-long
+    )
 
     _MOVE_CHART = """
             if (myVizzu_{c_id}.parentNode && myVizzu_{c_id}.parentNode.parentNode) {{
@@ -63,11 +45,14 @@ class DisplayTemplate:
             myVizzu_{id}.parentNode.insertBefore(myVizzu_{c_id}, myVizzu_{id});
             """
 
+    SCROLL = """myVizzu_{c_id}.scrollIntoView({{behavior: "auto", block: "center"}});"""
+
     ANIMATE = {
         DisplayTarget.BEGIN: _SCRIPT.format(
             """
             myVizzu_{id}.parentNode.parentNode.style.display = "none";
             chart_{c_id} = chart_{c_id}.then(chart => {{
+                {scroll}
                 return {animation};
             }});
             """
@@ -78,6 +63,7 @@ class DisplayTemplate:
             myVizzu_{{id}}.parentNode.parentNode.style.display = "none";
             chart_{{c_id}} = chart_{{c_id}}.then(chart => {{{{
                 {indent(_MOVE_CHART, "    ", lambda line: True)}
+                {{scroll}}
                 return {{animation}};
             }}}});
             """
@@ -88,6 +74,7 @@ class DisplayTemplate:
             myVizzu_{{id}}.parentNode.parentNode.style.display = "none";
             {_MOVE_CHART}
             chart_{{c_id}} = chart_{{c_id}}.then(chart => {{{{
+                {{scroll}}
                 return {{animation}};
             }}}});
             """
@@ -351,9 +338,10 @@ class Chart:
         self._div_width = width
         self._div_height = height
         self._display_target = DisplayTarget(display)
+        self._scroll_into_view = True
 
         self._display(
-            DisplayTemplate.INIT[self._display_target].format(
+            DisplayTemplate.INIT.format(
                 id=uuid.uuid4().hex[:7],
                 c_id=self._c_id,
                 vizzu=self._vizzu,
@@ -361,6 +349,14 @@ class Chart:
                 div_height=self._div_height,
             )
         )
+
+    @property
+    def scroll_into_view(self):
+        return self._scroll_into_view
+
+    @scroll_into_view.setter
+    def scroll_into_view(self, scroll_into_view):
+        self._scroll_into_view = bool(scroll_into_view)
 
     def feature(self, name, value):
         feature = Feature(name, value).dump()
@@ -379,9 +375,19 @@ class Chart:
 
         animation = self._merge_animations(animations)
         animation = Animate(animation, options).dump()
+
+        scroll = (
+            DisplayTemplate.SCROLL.format(c_id=self._c_id)
+            if self._scroll_into_view
+            else ""
+        )
+
         self._display(
             DisplayTemplate.ANIMATE[self._display_target].format(
-                id=uuid.uuid4().hex[:7], c_id=self._c_id, animation=animation
+                id=uuid.uuid4().hex[:7],
+                c_id=self._c_id,
+                animation=animation,
+                scroll=scroll,
             )
         )